Make the token optional
